LinguaGO-Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
groq
langgraph
langchain-core
langchain-groq
supabase
